qat: add kernel mode plugin
