+++ cmd/qat_plugin/kerneldrv/test_data/all_is_good/c6xx_dev2.conf
