+++ cmd/qat_plugin/kerneldrv/test_data/cant_parse_num_processes/c6xx_dev0.conf
@@ -1,0 +1,5 @@
+[SSL]
+NumberCyInstances = 6
+NumberDcInstances = 2
+NumProcesses = this is error
+LimitDevAccess = 0
