+++ cmd/qat_plugin/kerneldrv/test_data/missing_pinned_section/c6xx_dev1.conf
@@ -1,0 +1,193 @@
+################################################################
+# This file is provided under a dual BSD/GPLv2 license.  When using or
+#   redistributing this file, you may do so under either license.
+# 
+#   GPL LICENSE SUMMARY
+# 
+#   Copyright(c) 2007-2018 Intel Corporation. All rights reserved.
+# 
+#   This program is free software; you can redistribute it and/or modify
+#   it under the terms of version 2 of the GNU General Public License as
+#   published by the Free Software Foundation.
+# 
+#   This program is distributed in the hope that it will be useful, but
+#   WITHOUT ANY WARRANTY; without even the implied warranty of
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+#   General Public License for more details.
+# 
+#   You should have received a copy of the GNU General Public License
+#   along with this program; if not, write to the Free Software
+#   Foundation, Inc., 51 Franklin St - Fifth Floor, Boston, MA 02110-1301 USA.
+#   The full GNU General Public License is included in this distribution
+#   in the file called LICENSE.GPL.
+# 
+#   Contact Information:
+#   Intel Corporation
+# 
+#   BSD LICENSE
+# 
+#   Copyright(c) 2007-2018 Intel Corporation. All rights reserved.
+#   All rights reserved.
+# 
+#   Redistribution and use in source and binary forms, with or without
+#   modification, are permitted provided that the following conditions
+#   are met:
+# 
+#     * Redistributions of source code must retain the above copyright
+#       notice, this list of conditions and the following disclaimer.
+#     * Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in
+#       the documentation and/or other materials provided with the
+#       distribution.
+#     * Neither the name of Intel Corporation nor the names of its
+#       contributors may be used to endorse or promote products derived
+#       from this software without specific prior written permission.
+# 
+#   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+#   "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+#   LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+#   A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#   OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+#   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+#   DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+#   THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+#   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+#   OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+# 
+# 
+#  version: QAT1.7.L.4.2.0-00012
+################################################################
+[GENERAL]
+ServicesEnabled = cy;dc
+
+ConfigVersion = 2
+
+#Default values for number of concurrent requests*/
+CyNumConcurrentSymRequests = 512
+CyNumConcurrentAsymRequests = 64
+
+#Statistics, valid values: 1,0
+statsGeneral = 1
+statsDh = 1
+statsDrbg = 1
+statsDsa = 1
+statsEcc = 1
+statsKeyGen = 1
+statsDc = 1
+statsLn = 1
+statsPrime = 1
+statsRsa = 1
+statsSym = 1
+KptEnabled = 0
+
+# Disable public key crypto and prime number
+# services by specifying a value of 1 (default is 0)
+PkeServiceDisabled = 0
+
+# Specify size of intermediate buffers for which to
+# allocate on-chip buffers. Legal values are 32 and
+# 64 (default is 64). Specify 32 to optimize for
+# compressing buffers <=32KB in size.
+DcIntermediateBufferSizeInKB = 64
+
+# This flag is to enable device auto reset on heartbeat error
+AutoResetOnError = 0
+
+##############################################
+# Kernel Instances Section
+##############################################
+[KERNEL]
+NumberCyInstances = 1
+NumberDcInstances = 1
+
+# Crypto - Kernel instance #0
+Cy0Name = "IPSec0"
+Cy0IsPolled = 0
+Cy0CoreAffinity = 0
+
+# Data Compression - Kernel instance #0
+Dc0Name = "IPComp0"
+Dc0IsPolled = 0
+Dc0CoreAffinity = 0
+
+##############################################
+# User Process Instance Section
+##############################################
+[SSL]
+NumberCyInstances = 6
+NumberDcInstances = 2
+NumProcesses = 1
+LimitDevAccess = 0
+
+# Crypto - User instance #0
+Cy0Name = "SSL0"
+Cy0IsPolled = 1
+# List of core affinities
+Cy0CoreAffinity = 9
+
+# Crypto - User instance #1
+Cy1Name = "SSL1"
+Cy1IsPolled = 1
+# List of core affinities
+Cy1CoreAffinity = 10
+
+# Crypto - User instance #2
+Cy2Name = "SSL2"
+Cy2IsPolled = 1
+# List of core affinities
+Cy2CoreAffinity = 11
+
+# Crypto - User instance #3
+Cy3Name = "SSL3"
+Cy3IsPolled = 1
+# List of core affinities
+Cy3CoreAffinity = 12
+
+# Crypto - User instance #4
+Cy4Name = "SSL4"
+Cy4IsPolled = 1
+# List of core affinities
+Cy4CoreAffinity = 13
+
+# Crypto - User instance #5
+Cy5Name = "SSL5"
+Cy5IsPolled = 1
+# List of core affinities
+Cy5CoreAffinity = 14
+
+
+# Data Compression - User instance #0
+Dc0Name = "Dc0"
+Dc0IsPolled = 1
+# List of core affinities
+Dc0CoreAffinity = 9
+
+# Data Compression - User instance #1
+Dc1Name = "Dc1"
+Dc1IsPolled = 1
+# List of core affinities
+Dc1CoreAffinity = 10
+
+[SHIM]
+NumberCyInstances = 1
+NumberDcInstances = 0
+NumProcesses = 2
+LimitDevAccess = 0
+
+# Crypto - User instance #0
+Cy0Name = "UserCY0"
+Cy0IsPolled = 1
+# List of core affinities
+Cy0CoreAffinity = 0
+
+[PINNED]
+NumberCyInstances = 1
+NumberDcInstances = 0
+NumProcesses = 4
+LimitDevAccess = 1
+# Crypto - User instance #0
+Cy0Name = "UserCY0"
+Cy0IsPolled = 1
+# List of core affinities
+Cy0CoreAffinity = 0
